% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{atlastrack}
\alias{atlastrack}
\title{AtlasTrack Fiber Tract Atlas}
\usage{
atlastrack()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object.
}
\description{
Brain atlas for the AtlasTrack probabilistic white matter fiber tract
parcellation, commonly used with ABCD study data.
}
\examples{
atlastrack()

library(ggseg.formats)
plot(atlastrack())
}
\references{
Hagler DJ Jr, et al. (2019). "Image processing and analysis
methods for the Adolescent Brain Cognitive Development Study."
\emph{NeuroImage}, 202:116091.
\doi{10.1016/j.neuroimage.2019.116091}
}
\concept{ggseg_atlases}
\concept{tract_atlases}
