% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegAtlastrack-package.R
\docType{package}
\name{ggsegAtlastrack-package}
\alias{ggsegAtlastrack}
\alias{ggsegAtlastrack-package}
\title{ggsegAtlastrack: AtlasTrack Fiber Tract Atlas for the 'ggseg' Ecosystem}
\description{
AtlasTrack probabilistic white matter fiber tract atlas for the 'ggseg' ecosystem. Provides 'ggseg_atlas' objects for DTI fiber tract visualization, commonly used with ABCD study data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegAtlastrack}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegAtlastrack/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Pineiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
